educational_platform/requirements.txt
asgiref==3.7.2
Django==4.0
djangorestframework==3.14.0
pip==23.2.1
pytz==2023.3.post1
setuptools==63.2.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
